Reverse sorting order
Hi there, thanks for this awesome and cleanly written tool!

I've been using this for a few days now and realised after a while that the order in which it outputs email addresses seemed inversely correlated to how many emails I had exchanged with those addresses. Changing `sort_by_frequency` as in this patch made things work as expected for me.